Refactored Difficulty Colors
