+++ src/lib/Color_Config.json
@@ -1,0 +1,27 @@
+{
+    "Subject_color":  {
+         "Deutsch":"#B52020",
+         "Mathe":"#202BB5",
+         "Informatik":"#303030",
+         "Geschichte":"#F45E00",
+         "Politik":"#B520AB",
+         "Physik":"#707070",
+         "Biologie":"#31AD03",
+         "Religion":"#08B4CB",
+         "Chemie":"#FAE700"
+     },
+    "Difficulty_color": {
+      "Leicht": {
+        "bg": "rgba(25, 186, 63, 0.3)",
+        "border": "rgba(25, 186, 63, 1)"
+      },
+      "Mittel": {
+        "bg": "rgba(255, 186, 0, 0.3)",
+        "border": "rgba(255, 186, 0, 1)"
+      },
+      "Schwer": {
+        "bg": "rgba(186, 25, 25, 0.3)",
+        "border": "rgba(186, 25, 25, 1)"
+      }
+    }
+ }
