--- src/lib/Subjekt_Config.json
@@ -1,14 +1,0 @@
-{
-    "Subject_color":  {
-         "Deutsch":"#B52020",
-         "Mathe":"#202BB5",
-         "Informatik":"#303030",
-         "Geschichte":"#F45E00",
-         "Politik":"#B520AB",
-         "Physik":"#707070",
-         "Biologie":"#31AD03",
-         "Religion":"#08B4CB",
-         "Chemie":"#FAE700"
-     }
- }
-
